backend/python/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Numerical Computing & Financial Analysis (100% Accuracy)
numpy==1.26.2
pandas==2.1.4
scipy==1.11.4
scikit-learn==1.3.2
statsmodels==0.14.0

# Financial Calculations
yfinance==0.2.28
quantlib==1.32
py_vollib==1.0.1
empyrical==0.5.5

# Performance & Optimization
numba==0.58.1
cython==3.0.6

# Data Validation & Quality
marshmallow==3.20.1
cerberus==1.3.5

# HTTP & API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Caching & Database
redis==5.0.1
asyncpg==0.29.0
sqlalchemy==2.0.23

# Testing & Validation
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
